fix mysql config

+++ cookbooks/mysql/recipes/default.rb
@@ -11,7 +11,7 @@ execute "change mysql root password" do
   only_if "echo 'select 1;' | mysql --user=root --password= "
 end
 
-unless node.mysql[:keep_test]
+if node[:mysql] && node.mysql[:keep_test] != true
   execute "remove test mysql database" do
     command "echo 'drop database test;' | mysql --user=root --password=#{root_mysql_password} "
     only_if "echo 'show databases;' | mysql --user=root --password=#{root_mysql_password} | grep ^test$ "
